fix warnings

+++ ExoEngine/src/ExoEngine/ECS/Components/Attributes.h
@@ -47,7 +47,7 @@ namespace EM
 		float mFiringCoolDown;
 		float mFireDurationTimer;
 		float mChargeCoolDown;
-		float mHazardDmg;
+		int mHazardDmg;
 		HazardTypes mHazardType;
 	};
 }
